Brings non-recursive session merging back for `request_cached`
